Move everything to new file but keep it working
